+++ source/userCmds.h
@@ -52,10 +52,10 @@ enum class DialogTypes {
    e.) "menuItemB2" (hierarchical ID = {1, 2, 1})
  */
 struct userMenuInfo {
-    QString          umiName;          // hierarchical name of menu item (w.o. language mode info)
-    bool             umiIsDefault;     // menu item is default one ("@*")
-    std::vector<int> umiLanguageModes; // list of applicable lang. modes
-    size_t           umiDefaultIndex;  // array index of menu item to be used as default, if no lang. mode matches
+    QString             umiName;          // hierarchical name of menu item (w.o. language mode info)
+    bool                umiIsDefault;     // menu item is default one ("@*")
+    std::vector<size_t> umiLanguageModes; // list of applicable lang. modes
+    size_t              umiDefaultIndex;  // array index of menu item to be used as default, if no lang. mode matches
 };
 
 struct MenuData {
